* modules/codec/ffmpeg/encoder.c: hmm better be safe than sorry.

+++ modules/codec/ffmpeg/encoder.c
@@ -352,7 +352,7 @@ int E_(OpenEncoder)( vlc_object_t *p_this )
         av_reduce( &p_context->sample_aspect_ratio.num,
                    &p_context->sample_aspect_ratio.den,
                    i_aspect_num * (int64_t)p_context->height,
-                   i_aspect_den * p_context->width, 1 << 30 );
+                   i_aspect_den * (int64_t)p_context->width, 1 << 30 );
 #else
         p_context->aspect_ratio = ((float)p_enc->fmt_in.video.i_aspect) /
             VOUT_ASPECT_FACTOR;
